Move KernelBuilder into its own file
